[fix] metrics_handlers: castear mrr::float8 para compatibilidad con pgx v5 y numeric(10,2)

+++ internal/delivery/http/api/metrics_handlers.go
@@ -30,7 +30,7 @@ func HandleAdminMetrics(w http.ResponseWriter, r *http.Request) {
 	}
 	err := deps.DB.QueryRow(r.Context(), `
 		SELECT calculated_at, total_accounts, active_accounts, suspended_accounts,
-		       mrr, total_patients, total_notes, total_allergies, total_prescriptions
+		       mrr::float8, total_patients, total_notes, total_allergies, total_prescriptions
 		FROM metrics_snapshot
 		ORDER BY calculated_at DESC
 		LIMIT 1
